requirements.txt
streamlit
pandas
numpy
polygon-api-client
matplotlib